refactor: split CTA into separate components
This way we can keep the React components with fetching logic by
themselves. Also added some loading functionality to the form signup.
