+++ src/pages/index.astro
@@ -1,5 +1,5 @@
 ---
-import { CTA } from "../components/CTA";
+import CTA from "../components/CTA.astro";
 import ImageGrid from "../components/ImageGrid.astro";
 import Heading from "../components/ui/Heading.astro";
 import BaseLayout from "../layouts/BaseLayout.astro";
@@ -21,7 +21,7 @@ import BaseLayout from "../layouts/BaseLayout.astro";
                 Frustrated? Let’s break through the plateau 💪
               </p>
             </div>
-            <CTA client:visible />
+            <CTA />
           </div>
         </div>
       </div>
